cleanup node utils
