+++ src/router.ts
@@ -1,12 +1,12 @@
 import { createRouter as _createRouter } from 'radix3'
-import type { HTTPMethod, EventHandler, DeprecatedHandler } from './types'
+import type { HTTPMethod, EventHandler } from './types'
 import { createError } from './error'
 import { eventHandler, toEventHandler } from './event'
 
 export type RouterMethod = Lowercase<HTTPMethod>
 const RouterMethods: RouterMethod[] = ['connect', 'delete', 'get', 'head', 'options', 'post', 'put', 'trace', 'patch']
 
-export type RouterUse = (path: string, handler: EventHandler | DeprecatedHandler, method?: RouterMethod | RouterMethod[]) => Router
+export type RouterUse = (path: string, handler: EventHandler, method?: RouterMethod | RouterMethod[]) => Router
 export type AddRouteShortcuts = Record<RouterMethod, RouterUse>
 
 export interface Router extends AddRouteShortcuts {
