+++ src/utils/index.ts
@@ -1,3 +1,4 @@
+export * from './route'
 export * from './body'
 export * from './cache'
 export * from './consts'
